Setting default graph via URL parameter
SPARQL 1 and 1.1 both know the URL paramter 'default-graph-uri'. It can
is used to set the default graph of a query. Previously, when doing a
query on a context of a Conjunctive graph backed by the SPARQLStore, a
GRAPH clause is injected into the INSERT. Modifying the query is not
robust. Now, the default-graph-uri is set while leaving the query
unmodified.

The same is possible in SPARQLUpdateStore with the parameter
using-graph-uri, but it is not used by this commit, because it does
not work with my SPARQL endpoint.