Unpin html5lib 0.95 requirement
Due to an html5lib regression (described in the thread at
https://groups.google.com/d/msg/rdflib-dev/ZcAgKzhS3vI/3mxIJz4rwWUJ) , we
opened https://github.com/html5lib/html5lib-python/issues/67 on June 17th, 2013
and pinned html5lib to 0.95 in setup.py. The bug was fixed and a new release of
html5lib (1.0b3) was cut on July 24, 2013. The current version of html5lib is
0.999; let's unpin that html5lib requirement.

Signed-off-by: Dan Scott <dan@coffeecode.net>