Improved update rewriting
In order to find beginnings and ends of blocks, a very simplified
grammar is used. This has the advantage that extensions of the SPARQL
language may work. On the other hand, it may happen that the query
rewriting produces invalid output. The complete SPARQL Update grammar[1]
(including strings containing { and IRIs containing ') should be
supported, given my reasonings are correct.

See #395

[1]: http://www.w3.org/TR/sparql11-query/#sparqlGrammar