make external_graph_libs tests optional (only run if ext lib is installed)
networkx shouldn't be a dependency of rdflib just to run tests. Also
graph_tool heavily depends on C libs and can't easily be installed via pip
(or via apt-get on travis).
Hence, the doctests are duplicated into proper tests that auto skip if
networkx or graph_tool can't be imported.