Prevent RDFa parser from failing on time elements with child nodes
The _get_literal() function in the RDFa parser referred to a self instance that
does not exist. This led to a NameError exception if the HTML to be parsed
included a TIME element with one or more child nodes, such as
'<time><em>2016-01-01</em></time>'. Removing the reference to 'self' fixes
the problem and fixes #576.

Signed-off-by: Dan Scott <dan@coffeecode.net>