namespace.py fix compute_qname missing namespaces
The way the compute_qname worked was to call split_uri
and then see if there was a match in self.store.prefix.
This produced incomplete behavior for namesapces that
end in LETTERS_ instead of / or #. This commit corrects
this behavior by iterating through name and testing
namespace + name[:i] to see if there is a matching prefix.

namespace.py fix to allow numbers to start names