term.py added _XSD_TIME to _NO_TOTAL_ORDER_TYPES
gYearMonth and gYear (as well as the other g* xsd types) shoud also
be included, but their current implementation does not match the xsd
spec so they are commented out. If the g* types are implemented as
durations or time intervals sorting will likely be a much more
complicated issue.