namespace.py serializers/rdfxml.py stric qnames
The rdfxml serializer relied on namespace.py compute_qname to enforce
qname correctness. Chagnging split characteristics to support n3
prefixes broke that assumption. This commit is a first pass at fixing
the behvior. It may be worth renaming functions to make it clear that
the current 'compute_qname' function is no longer actually computing
a qname but a n3 prefix (or something like that).