added a class level option to roundtrip prefixes
This is the simplest way to allow configuration of whether prefixes are
roundtripped during serialization. If someone needs to roundtrip the
prefixes then they can import the serializer into their code and see the
class variable, or if they care enough about the exact way things are
serialized they will likely be subclassing the serializer anyway.