Do not use current time in sparql TIMEZONE
Calculating the timezone should use date date and time from the
datetime object the timezone is requested from instead from the
current time. This makes a difference if the timezone for the given
location has changed in the past. In this particular case it is probably
not possible as iso 8601 seems not have a method to specify the
time zone via a location.