Fix SPARQL update parsing to handle arbitrary amounts of triples in inserts
pyparsing handling of recursive rules is limited by python's recursion limit
and the rule that should handle multiple triples in data inserts is
recursive.

The consequence of this is that the amount of triples that can be
inserted becomes limited by python's recursion limit.

To address this the rule has been rewritten to not be recursive.

@rchateauneu thanks for the help with rule rewriting.