Fix shield for CI status
The old shield was still pointing to travis, this patch changes
it to use the Drone CI shield shield.

The link is to the status for all branches, I can't find a link
specifically to the master branch status, the links from the branch
status page goes to individual build numbers.