Compile docs in GitHub Actions CI
This is to make sure that there are no doc build failures introduced by
PRs.

In future I want to expand this check to be more strict. Since it comes up
separately in CI it can be ignored on reviwer discrescion but at least
it will make problems visible.

Also:
- Fix some things in docs that were causing warnings. There are still
  many warnings related to references but I will fix them later.
- Added a `docs` evnrionment for tox which builds the docs, this is also
  what is used from CI.