Add IdentifiedNode abstract intermediary class
This patch adds and exposes an intermediary class `IdentifiedNode` as a
superclass of `URIRef` and `BNode`.  From review of the subclass methods
for identical implementations, two appeared to be able to move into this
superclass.

Thanks to Nicholas Car for finding this pragmatic name.

This patch addresses at least some of Issue 1526.

References:
* https://github.com/RDFLib/rdflib/issues/1526

Cc: Nicholas Car <nicholas.car@surroundaustralia.com>
Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>