Add typing to rdflib.namespace
Everything is as typed as it can be in here, there is maybe some
improvement that can be made to `__new__` in some places but I think
there may be some runtime changes needed there so I don't want to change
that now, the inferred return types are close enough.

Some of the type hints are based on pytest-monkeytype output which was
slightly tuned for simplicity and with some additional typing for methods
that are not being tested.

This changeset contains type changes only and does not change runtime
behaviour.