Apply IdentifiedNode to Graph iterators triples() et al.
This patch is known to be incomplete with respect to other needed
revisions to type signatures, both in Graph subclasses and in graph
consumers elsewhere under `/rdflib`.  However, running this change
through a type checker raised some questions on `Node` specificity and
circular type dependencies, and so is posted for discussion.

References:
* https://github.com/RDFLib/rdflib/issues/1696

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>