Fix turtle serialization of PNames that contain brackets
Brackets in pnames will now be escaped in output.

This fix is based on a suggestion by Niklas Lindström (@niklasl).

This is somewhat of a stopgap fix, there are some problems being
masked and various cases that could still result in questionable Turtle being
written out, however these can be addressed seperately.