Use the `property` built-in as a decorator
The use of the `property` built-in as a function does not allow for
the same type annotations as using `property` as a decorator.

This change should not affect runtime behaviour, but I added tests to
cover all properties that were changed to decorators explicitly just to
make sure they work as they should.

There are still other places where the `property` built-in is used as a
function but this is outside of the core parts of RDFLib.