Fix type checking errors
Type checking passes now.

I also included some changes which reduces
the footprint of this change so that it becomes almost entirely just
change of type annotations.

This may be a bit on the pendatic side, and I would be fine if you keep
the changes as they were, but I would rather keep refactoring seperate from
changes to add typing.