TypeErrors from Results do not propagate through list creation
Creating a list from from a query result that generates a `TypeError`
succeeds and returns an empty list.

The correct behaviour would be that the TypeError propagates through the
list constuctor and that no list is created so that the caller can be
aware that an error occured and that the caller does not get a seemingly
okay list object which does not contain the result of the executed
query.

This patch adds some tests for SPARQL, including expected failure tests
that demonstrate the problem of TypeErrors not propagating through
list constructor.