Add two xfails related to Example 2 from RDF 1.1 TriG specification
The first xfail occurs during round tripping, TriG seems to be making
some mistake when encoding blank nodes, as it is encoding that "Bob"
knows someone who does not exist. This was reported by @gjhiggins in
https://github.com/RDFLib/rdflib/pull/1796#discussion_r846781069

The second xfail seems to be related to hextuple parsing, when comparing
the hextuple parsed result of Example 2 with the TriG parsed
graph of Example 2 the graphs are not isomorphic more than 70% of the time, but
sometimes they are isomorphic. Inoticed this while adding the xfail for
the issue @gjhiggins noticed.

Other changes:
- Added `simple_quad` to variants tests with HexTuple and TriG format.
- Added an additional exact_match assert for variants which can be used
  to sidestep some of the known issues with isomorphic graph detection.
  This is useful for graphs with no BNodes.
- Also added round-tripping for `variants/simple_quad.trig`.
- Various changes to ensure determensitic ordering so that it is easier
  to compare things visually and so that tests always do the exact same
  thing in the exact same order.