test: Add tests for SPARQL parsing and serialization
This patch adds some tests for SPARQL serialization and deserialization
with checks based on round tripping where possible.

Some of these tests are expected failures in that they seem to pass fo
some parsers or serializers, but not others. If there is agreement on
what parameters should be supported we can fix the parsers and
serializers to actually support them.

There are also xfails for the handling of string paths on windows, as
this seems to fail for all serializers.