Fix handling of EXISTS inside BIND
This patch fixes an issue with `BIND( EXISTS ... )` in SPARQL,
for example:

```sparql
SELECT * WHERE {
    BIND(
	EXISTS {
	    <http://example.com/a>
	    <http://example.com/b>
	    <http://example.com/c>
	}
	AS ?bound
    )
}
```

The graph pattern of `EXISTS` needs to be translated for it to operate
correctly during evaluation, but this was not happening. This patch
corrects that so that the graph pattern is translated as part of
translating `BIND`.

This patch also adds a bunch of tests for EXISTS to ensure there is no
regression and that various EXISTS cases function correctly.

Fixes https://github.com/RDFLib/rdflib/issues/1472