Added Taskfile with development tasks
This patch adds a Taskfile with various tasks that is useful for
RDFLib development. Details of the available tasks can be seen from
`task help` output and some more information is included in the RDFLib
developers guide.

For more information about Taskfiles see https://taskfile.dev/#/usage.

The GitHub action pipeline was also changed to now use tasks from the
Taskfile which simplifies the pipeline quite a bit.

This patch also adds development container which can be used to run
various development task and which can also be used with development
container enabled tools such as VSCode or GitHub codespaces.

This is all being done in preperation for moving coveralls reporting to
the GitHub actions workflow.

Other changes:
- Removed the `Makefile` and associated `test/Dockerfile` as it is now replaced by a `Taskfile.yml`.