Enable and baseline flakeheaven
This patch adds flakeheaven with a baseline, the baseline ignores all
existing errors and only reports on new errors.

The benefit with this approach is that we can get full flake8 on new
changes without having to first address the flake8 warnings with our
existing codebase.

Baselined errors can still be seen with:
- `flakeheaven lint --baseline /dev/null`
- `flake8`

This also adds a flakeheaven to GitHub Actions in a way which will
result in independent reporting.

Other changes:
- Simplified caching behaviour in GitHub Actions
- Removed flake8-black because it does not work well

  https://github.com/peterjc/flake8-black/issues/59
- Remove redundant/duplicated config for flake8