Remove `pytest-subtest`
`unittest.TestCase.subTest` is a `unittest` construct that should be
avoided in favor of pytest parameterization.

`pytest-subtest` is a pytest plugin that adds support for
`unittest.TestCase.subTest` to pytest.

This plugin was needed in the past due to the use of `subTest`, but all uses of
`subTest` has been replaced with pytest parameterization recently.

The removal of this plugin is desirable as it bumps heads with other
plugins that are useful and generally `subTest` should be avoided due to
limitations in annotating this with skip of xfail.