Add test for expected behavior of sorted Turtle output
This patch adds a test to start specifying what sorting Turtle output
would look like.  This is intended to start discussion about
expectations of blank node sorting, and to set an initial interface for
triggering sorted output with a propagated keyword argument in
`Graph.serialize()`.

This patch will fail CI, but should not fail for code-style reasons.
The new test script was reviewed with black, flake8, isort, and
mypy (--strict).

References:
* https://github.com/RDFLib/rdflib/issues/1890

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>