chore: remove pre python 3.5 regex related workaround
Removing a pre-python 3.5 regex related workaround as it uses a private
method from the `re` module, and mypy complains about this, it is also
no longer needed as we no longer support pre 3.5 versions of python.

No tests are added as all changed parts of the `Builtin_REPLACE`
function is already covered by existing tests.