chore: rename default branch to `main`
We are still using a somewhat outmoded name for our default branch, this
PR captures the decision to change it as a decision record and prepares
the rest of the repo for this change.

Once merged, the main branch of RDFLib will be renamed using the
procedure documented [here](https://github.com/github/renaming).

Other changes:

- Add area for decision records (this may be merged in another PR before
  this one).
- Remove duplicated sections from `docs/upgrade4to5.rst` as they cause
  problems when using `sphinx.ext.autosectionlabel` and serve no
  purpose.