compute_qname handle case where name could be unbound
In NamespaceManager.compute_qname it is possible for the variable name
to be unbound when a namespace is used as a predicate e.g. ns2: and
the namespace itself cannot be split any further.

Fixing this usually just kicks the can down the road because the only
reason to hit this condition is that you are trying to serialize a
predicate that has no valid NCName form into an xml format that
requires one, making it impossible to complete the conversion.
That said, this fix should make it easier to understand what is going
on by unmasking the real issue.