namespace.get_longest_namespace avoid concurrent modification issues
When using async code that operates on namespaces it is possible for
the namespacies trie dict to be modified in one coro while it is being
iterated through in another, causing a RuntimeError. To avoid this we
switch to always iterate over trie.keys(). This is safe because any
concurrent modification will either have no impact or will duplicate
the operation leaving the trie in the same state.