docs: don't use kroki
The kroki server is currently experiencing some issues which breaks our
build, this change eliminates the use of kroki in favor of directly
using the generated svg images which is checked into git alongside the
plantuml sources.

I also added a task to the Taskfile to re-generate the svg images from
the plantuml sources by calling docker.