fix: restore the 6.1.1 default bound namespaces
The namespaces bound by default by `rdflib.graph.Graph` and
`rdflib.namespace.NamespaceManager` was reduced in version 6.2.0 of RDFLib,
however, this also would cause code that worked with 6.1.1 to break, so this
constituted a breaking change. This change restores the previous behavior,
binding the same namespaces as was bound in 6.1.1.

To bind a reduced set of namespaces the `bind_namespaces` parameter of
`rdflib.graph.Graph` or `rdflib.namespace.NamespaceManager` can be used.

- Closes <https://github.com/RDFLib/rdflib/issues/2103>.