docs: recommend making an issue before making an enhancement
Suggest that contributors first make an issue to get in principle
agreement for pull requests before making the pull request.

Enhancements can be controversial, and we may reject the enhancement
sometimes even if the code is good, as it may just not be deemed
important enough to increase the maintenance burden of RDFLib.

Other changes:
- Updated the checklist in the pull request template to be more accurate to
  current practice.
- Improved grammar and writing in the pull request template, contribution guide
  and developers guide.