Fix comment doesn't describe behavior
Comment refers to a person that knows bob and code would return a name, but this would only work if the triple
`person foaf:name bob .`
is part of the dataset. As this is a very uncommon way to model a `foaf:knows` the code was adjusted to match the description.